requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
Pillow>=10.0.0
qrcode>=7.4.2
urllib3>=2.0.0
